+++ packages/elasticsearch-plugin/index.ts
@@ -1,3 +1,4 @@
+export * from './src/adapter';
 export * from './src/plugin';
 export * from './src/options';
 export * from './src/types';