+++ packages/opensearch-plugin/src/constants.ts
@@ -1,3 +1,0 @@
-export const OPENSEARCH_OPTIONS = Symbol('OPENSEARCH_OPTIONS');
-export const VARIANT_INDEX_NAME = 'variants';
-export const loggerCtx = 'OpenSearchPlugin';